FollowToggleButton, MessageButton 컴포넌트를 다른 파일로 분리
